requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
httpx
pydantic[email]
python-jose[cryptography]
bcrypt
